STEP2 수정
