+++ CodeStarterCamp_Week4/error.swift
@@ -1,0 +1,13 @@
+//
+//  error.swift
+//  CodeStarterCamp_Week4
+//
+//  Created by hanjongwoo on 2022/07/15.
+//
+
+import Foundation
+
+enum RoutineError: Error {
+    case fatigueOverflow
+    case wrongInput
+}
